% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA Earthquake Data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{a data frame of NOAA earthquake.}
}
\value{
A \code{tibble} with a new column \code{Date} but removing
\code{Year}, \code{Mo} & \code{Dy} variables of the input data frame.
}
\description{
\code{eq_clean_data} takes NOAA earthquake dataset and does some data
 manipulations on it.
}
\details{
This function takes NOAA earthquake data and creates a \code{Date} column
 of type date by combining \code{Year}, \code{Mo} and \code{Dy}. It then
 makes sure that \code{Longitude} and \code{Latitude} are of type numeric.
}
\examples{
data(NOAA_data)
eq_clean_data(NOAA_data)

}
