% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean and Separate NOAA Earthquake Location Field}
\usage{
eq_location_clean(data)
}
\value{
a \code{tibble} with one additional column which is the result of the
 previous \code{Location Name} being divided into two columns \code{Country}
 and \code{Location}.
}
\description{
This function takes a NOAA earthquake data frame and do some data
manipulation on \code{Locaiton Name}.
}
\details{
\code{eq_location_clean} function manipulates the \code{Location Name} by
 separating country name and location name into \code{Country} and
 \code{Location} columns. It then turns \code{Location} into Title Case.
}
\examples{
data(NOAA_data)
eq2 <- eq_location_clean("NOAA_data")

}
