% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Leaflet Map Tools}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{The \code{\link{NOAA_data}} data frame.}

\item{annot_col}{The name of the column to be used for annotation and it
has to be a character strings of length one.}
}
\value{
Returns a \code{leaflet} map.
}
\description{
\code{eq_map} creates a \code{\link[leaflet]{leaflet}} map
showing the location of each earthquakes on the map.
}
\details{
The function maps the epicenters \code{Longitude} and \code{Latitude} and
annotates each point with in popup window containing annotation data stored
in a column of the data frame. The use can choose which column to be used
for popup.
}
\examples{
NOAA_data \%>\%
 eq_clean_data() \%>\%
 eq_location_clean() \%>\%
 dplyr::filter(Country == "MEXICO" & lubridate::year(Date) >= 2000) \%>\%
 eq_map(annot_col = "Date")

}
