% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom for creating a timeline from earthquake data}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetis mappings created by \code{\link[ggplot2]{aes}}.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{Override the default connection between \code{geom_timeline_label}
and stat_timeline_label(however I have not written a stat for this geom).}

\item{position}{Position adjustment, either as a string, or the result of a
call to position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends? NA,
the default, includes if any aesthetics are mapped. FALSE never includes,
and TRUE always includes. It can also be a named logical vector to finely
select the aesthetics to display.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning.
If TRUE, missing values are silently removed.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them.}

\item{...}{Other arguments passed on to \code{link[ggplot2]{layer}}.}
}
\description{
This function creates a new geom to create a timeline for a specified
data range and plots each earthquake as a point on that timline
}
\examples{
\dontrun{
ggplot(data = eq, aes(x = Date, y = Country,
color = Deaths, size = Mag)) + geom_timeline(alpha = 0.2)
}
}
