% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Raw NOAA Data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{data frame containing raw NOAA data}
}
\value{
a data frame of cleaned NOAA data
}
\description{
The function takes a raw NOAA data frame and returns a clean data frame.
}
\examples{
\dontrun{
eq_clean_data(raw_noaa_data)
}
}
