% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location Name}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{data frame containing raw NOAA data}
}
\value{
a data frame with clean location names
}
\description{
Cleans LOCATION_NAME by stripping out the country name (including the colon)
and converts names to title case (as opposed to all caps).
}
