% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Create interactive map with popup annotation}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{data frame with LONGITUDE, LATITUDE and
Mag columns}

\item{annot_col}{string used for popup annotation}
}
\value{
list or map
}
\description{
This function uses functions from \verb{leaflet package}
to create an interactive map with popup annotation.
The column to be used for annotation should be a
vector character or a string supplied through the
\code{annot_col} argument. Each row of the \code{data} supplied
is represented on the map with a \code{circle}  at LONGITUDE
and LATITUDE on that row. The radius of the circle is
proportional to the Mag on that row. The annotation is
displayed by clicking on the circle.
}
